flask-jwt-auth/requirements.txt
alembic
bcrypt
cffi
click
debugpy
coverage
Flask
Flask-Bcrypt
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-Testing
gunicorn
itsdangerous
Jinja2
Mako
MarkupSafe
psycopg2-binary==2.8.6 
ptvsd
pycparser
pyjwt==2.0.1
python-editor
six
SQLAlchemy
sqlalchemy-utils
Werkzeug
